refactoring: AuthenticationInterceptor 추상화
