+++ src/main/java/nextstep/subway/auth/ui/AuthenticationInterceptor.java
@@ -1,0 +1,51 @@
+package nextstep.subway.auth.ui;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import nextstep.subway.auth.domain.Authentication;
+import nextstep.subway.auth.domain.AuthenticationToken;
+import nextstep.subway.exception.InvalidPasswordException;
+import nextstep.subway.exception.UserDetailsNotExistException;
+import nextstep.subway.member.application.CustomUserDetailsService;
+import nextstep.subway.member.domain.LoginMember;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public abstract class AuthenticationInterceptor implements HandlerInterceptor {
+
+    private final AuthenticationConverter authenticationConverter;
+    private final CustomUserDetailsService customUserDetailsService;
+
+    public AuthenticationInterceptor(AuthenticationConverter authenticationConverter, CustomUserDetailsService customUserDetailsService) {
+        this.authenticationConverter = authenticationConverter;
+        this.customUserDetailsService = customUserDetailsService;
+    }
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
+        AuthenticationToken authenticationToken = authenticationConverter.convert(request);
+        Authentication authentication = authenticate(authenticationToken);
+        afterAuthentication(request, response, authentication);
+        return false;
+    }
+
+    public Authentication authenticate(AuthenticationToken authenticationToken) {
+        String principal = authenticationToken.getPrincipal();
+        LoginMember userDetails = customUserDetailsService.loadUserByUsername(principal);
+        checkAuthentication(userDetails, authenticationToken);
+        return new Authentication(userDetails);
+    }
+
+    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
+        if (userDetails == null) {
+            throw new UserDetailsNotExistException();
+        }
+        if (!userDetails.checkPassword(token.getCredentials())) {
+            throw new InvalidPasswordException();
+        }
+    }
+
+    public abstract void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException;
+}
