+++ src/main/java/nextstep/subway/auth/ui/token/TokenAuthenticationInterceptor.java
@@ -2,45 +2,33 @@ package nextstep.subway.auth.ui.token;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.io.CharStreams;
 import nextstep.subway.auth.domain.Authentication;
-import nextstep.subway.auth.domain.AuthenticationToken;
 import nextstep.subway.auth.dto.TokenResponse;
 import nextstep.subway.auth.infrastructure.JwtTokenProvider;
 import nextstep.subway.auth.ui.AuthenticationConverter;
-import nextstep.subway.exception.InvalidPasswordException;
-import nextstep.subway.exception.UserDetailsNotExistException;
+import nextstep.subway.auth.ui.AuthenticationInterceptor;
 import nextstep.subway.member.application.CustomUserDetailsService;
-import nextstep.subway.member.domain.LoginMember;
 import org.springframework.http.MediaType;
-import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.Map;
 
-public class TokenAuthenticationInterceptor implements HandlerInterceptor {
+public class TokenAuthenticationInterceptor extends AuthenticationInterceptor {
 
-    private final CustomUserDetailsService customUserDetailsService;
     private final JwtTokenProvider jwtTokenProvider;
-    private final AuthenticationConverter authenticationConverter;
     private final ObjectMapper mapper;
 
     public TokenAuthenticationInterceptor(CustomUserDetailsService customUserDetailsService, JwtTokenProvider jwtTokenProvider, AuthenticationConverter authenticationConverter) {
-        this.customUserDetailsService = customUserDetailsService;
+        super(authenticationConverter, customUserDetailsService);
         this.jwtTokenProvider = jwtTokenProvider;
-        this.authenticationConverter = authenticationConverter;
         this.mapper = new ObjectMapper();
     }
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
-        AuthenticationToken authenticationToken = authenticationConverter.convert(request);
-        Authentication authentication = authenticate(authenticationToken);
+    public void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
         TokenResponse tokenResponse = create(authentication);
         send(response, tokenResponse);
-        return false;
     }
 
     private void send(HttpServletResponse response, TokenResponse tokenResponse) throws IOException {
@@ -55,22 +43,5 @@ public class TokenAuthenticationInterceptor implements HandlerInterceptor {
         String token = jwtTokenProvider.createToken(payload);
         TokenResponse tokenResponse = new TokenResponse(token);
         return tokenResponse;
-    }
-
-    public Authentication authenticate(AuthenticationToken authenticationToken) {
-        String principal = authenticationToken.getPrincipal();
-        LoginMember userDetails = customUserDetailsService.loadUserByUsername(principal);
-        checkAuthentication(userDetails, authenticationToken);
-
-        return new Authentication(userDetails);
-    }
-
-    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
-        if (userDetails == null) {
-            throw new UserDetailsNotExistException();
-        }
-        if (!userDetails.checkPassword(token.getCredentials())) {
-            throw new InvalidPasswordException();
-        }
     }
 }
